refactor(db): Extract migration and patch modules from database.rs
Extract two modules from the monolithic database.rs (3,923 lines):

- migration.rs (708 lines): Lens migration functions
  - lens_store(), set_migration(), set_migration_in_txn()
  - maybe_reindex_after_migration(), has_migration()
  - Helper functions for placeholder creation

- patch.rs (1,748 lines): JSON Patch (RFC 6902) operations
  - patch_collection() and all helper methods
  - generate_patch_version_id()
  - JSON pointer operations (set, get, remove)
  - Field name substitution and validation

database.rs reduced from 3,923 to 1,537 lines (61% reduction).

All tests pass:
- cargo test -p db: 24 tests passed
- ffi-test run mutation/update: 86 passed, 0 failed
- ffi-test run query/simple: 434 passed, 1 failed (pre-existing)
