+++ crates/db/src/database.rs
@@ -3,25 +3,20 @@
 /// The DB struct is the main entry point for DefraDB operations.
 /// It manages the root store, creates transactions, and provides
 /// access to collections.
-use crate::collection::{collection_short_id, Collection};
+use crate::collection::Collection;
 use crate::collection_name::CollectionName;
 use crate::collection_snapshot::CollectionSnapshot;
 use crate::error::{Error, Result};
-use crate::index_manager::IndexManager;
-use crate::schema_loader::get_collections_by_collection_id;
 use crate::txn::DbTxn;
 use datastore::BasicTxn;
 use events::Bus;
 use identity::{Identity, RawIdentity};
 #[cfg(not(feature = "native"))]
 use lens::MemoryTransformStore;
+use lens::TransformStore;
 #[cfg(feature = "native")]
 use lens::WasmTransformStore;
-use lens::{
-    build_targeted_history, CollectionHistoryLink, Lens, LensConfig, LensDoc, TransformId,
-    TransformStore, DOC_ID_FIELD,
-};
-use schema::{CollectionSource, CollectionVersion, FieldKind, ScalarKind, ORPHAN_COLLECTION_ID};
+use schema::{CollectionSource, CollectionVersion};
 use std::collections::HashMap;
 use std::sync::atomic::{AtomicU64, Ordering};
 use std::sync::{Arc, RwLock};
@@ -100,20 +95,20 @@ impl DbOptions {
 /// This matches Go's DB struct in internal/db/db.go.
 pub struct DB<S: Store> {
     /// The underlying store.
-    store: Arc<S>,
+    pub(crate) store: Arc<S>,
     /// Options for this database instance.
     options: DbOptions,
     /// Counter for generating unique transaction IDs.
     txn_id_counter: AtomicU64,
     /// In-memory collection cache (name -> Collection).
-    collections: RwLock<HashMap<String, Collection>>,
+    pub(crate) collections: RwLock<HashMap<String, Collection>>,
     /// Event bus for subscription notifications.
     event_bus: Option<Arc<dyn Bus>>,
     /// Lens transform store for schema migrations.
-    lens_store: Arc<dyn TransformStore>,
+    pub(crate) lens_store: Arc<dyn TransformStore>,
     /// Pending migrations registered before their destination version exists.
     /// Maps dest_version_id -> (source_version_id, transform_id_string).
-    pending_migrations: RwLock<HashMap<String, (String, String)>>,
+    pub(crate) pending_migrations: RwLock<HashMap<String, (String, String)>>,
 }
 
 impl<S: Store> DB<S> {
@@ -340,584 +335,8 @@ impl<S: Store> DB<S> {
         self.txn_id_counter.load(Ordering::SeqCst)
     }
 
-    // =========================================================================
-    // Lens Migration Methods
-    // =========================================================================
-
-    /// Get a reference to the lens transform store.
-    ///
-    /// The lens store manages schema migration transforms that can be applied
-    /// when documents are fetched from older schema versions.
-    pub fn lens_store(&self) -> &Arc<dyn TransformStore> {
-        &self.lens_store
-    }
-
-    /// Set a migration between two schema versions.
-    ///
-    /// This registers a lens transform that will be applied to documents
-    /// when migrating from the source schema version to the destination.
-    ///
-    /// # Arguments
-    ///
-    /// * `config` - The lens configuration containing source/destination versions and transform
-    ///
-    /// # Returns
-    ///
-    /// The transform ID that was registered.
-    pub async fn set_migration(&self, config: LensConfig) -> Result<TransformId> {
-        let dest_version_id = config.destination_schema_version_id.clone();
-        let source_version_id = config.source_schema_version_id.clone();
-
-        let txn = self.new_txn(false).await?;
-
-        // Look up source and destination versions, creating placeholders if needed
-        // (matches Go's setMigration in internal/db/lens.go)
-        let (source_col, mut dst_col) = {
-            let systemstore = txn.systemstore()?;
-
-            // Look up source version
-            let src_key = CollectionKey::new(&source_version_id);
-            let src_data = systemstore
-                .get(&src_key.bytes())
-                .await
-                .map_err(Error::Storage)?;
-            let source_col: CollectionVersion = match src_data {
-                Some(data) => serde_json::from_slice(&data).map_err(|e| {
-                    Error::Serialization(format!(
-                        "failed to deserialize source schema '{}': {}",
-                        source_version_id, e
-                    ))
-                })?,
-                None => {
-                    // Source doesn't exist — create a placeholder
-                    let mut placeholder = CollectionVersion {
-                        version_id: source_version_id.clone(),
-                        collection_id: ORPHAN_COLLECTION_ID.to_string(),
-                        is_materialized: true,
-                        is_placeholder: true,
-                        ..CollectionVersion::new("", "", "", Vec::new())
-                    };
-                    placeholder.is_active = false;
-                    let data = serde_json::to_vec(&placeholder).map_err(|e| {
-                        Error::Serialization(format!(
-                            "failed to serialize source placeholder '{}': {}",
-                            source_version_id, e
-                        ))
-                    })?;
-                    systemstore
-                        .set(&src_key.bytes(), &data)
-                        .await
-                        .map_err(Error::Storage)?;
-                    placeholder
-                }
-            };
-
-            // Look up destination version
-            let dst_key = CollectionKey::new(&dest_version_id);
-            let dst_data = systemstore
-                .get(&dst_key.bytes())
-                .await
-                .map_err(Error::Storage)?;
-            let dst_col: CollectionVersion = match dst_data {
-                Some(data) => serde_json::from_slice(&data).map_err(|e| {
-                    Error::Serialization(format!(
-                        "failed to deserialize destination schema '{}': {}",
-                        dest_version_id, e
-                    ))
-                })?,
-                None => {
-                    // Destination doesn't exist — create a placeholder
-                    let mut placeholder = CollectionVersion {
-                        name: source_col.name.clone(),
-                        version_id: dest_version_id.clone(),
-                        collection_id: source_col.collection_id.clone(),
-                        is_materialized: true,
-                        is_placeholder: true,
-                        ..CollectionVersion::new("", "", "", Vec::new())
-                    };
-                    placeholder.is_active = false;
-                    // Store destination placeholder (same as source placeholder above)
-                    let data = serde_json::to_vec(&placeholder).map_err(|e| {
-                        Error::Serialization(format!(
-                            "failed to serialize destination placeholder '{}': {}",
-                            dest_version_id, e
-                        ))
-                    })?;
-                    systemstore
-                        .set(&dst_key.bytes(), &data)
-                        .await
-                        .map_err(Error::Storage)?;
-                    placeholder
-                }
-            };
-
-            (source_col, dst_col)
-        };
-
-        // Validate version adjacency
-        if let Some(ref prev) = dst_col.previous_version {
-            if prev.source_collection_id != source_col.version_id {
-                return Err(Error::InvalidPatch(format!(
-                    "cannot migrate between non-adjacent collection versions. \
-                     Destination '{}' already has previous version '{}', but migration source is '{}'",
-                    dest_version_id, prev.source_collection_id, source_version_id
-                )));
-            }
-        }
-
-        // Register the transform in the lens store
-        let transform_id = self
-            .lens_store
-            .add(config)
-            .await
-            .map_err(|e| Error::Lens(e.to_string()))?;
-
-        // Set the destination's previous_version with source and transform
-        dst_col.previous_version = Some(CollectionSource {
-            source_collection_id: source_col.version_id.clone(),
-            transform: Some(transform_id.to_string()),
-        });
-
-        tracing::debug!(
-            dest_version_id = %dest_version_id,
-            source_version_id = %source_version_id,
-            is_placeholder = dst_col.is_placeholder,
-            transform_id = %transform_id,
-            "set_migration: storing destination version with transform"
-        );
-
-        // Save the destination version
-        let collection_name = dst_col.name.clone();
-        let dst_key = CollectionKey::new(&dest_version_id);
-        let dst_data = serde_json::to_vec(&dst_col).map_err(|e| {
-            Error::Serialization(format!(
-                "failed to serialize destination schema '{}': {}",
-                dest_version_id, e
-            ))
-        })?;
-
-        {
-            let systemstore = txn.systemstore()?;
-            systemstore
-                .set(&dst_key.bytes(), &dst_data)
-                .await
-                .map_err(Error::Storage)?;
-
-            // Write CollectionVersionKey entries so get_collection_version_ids() can
-            // find these versions via prefix scan on /collection/version/{collection_id}/
-            if !source_col.collection_id.is_empty() {
-                let src_version_key =
-                    CollectionVersionKey::new(&source_col.collection_id, &source_version_id);
-                systemstore
-                    .set(&src_version_key.bytes(), b"1")
-                    .await
-                    .map_err(Error::Storage)?;
-            }
-            if !dst_col.collection_id.is_empty() {
-                let dst_version_key =
-                    CollectionVersionKey::new(&dst_col.collection_id, &dest_version_id);
-                systemstore
-                    .set(&dst_version_key.bytes(), b"1")
-                    .await
-                    .map_err(Error::Storage)?;
-            }
-        }
-        txn.commit().await?;
-
-        // Update in-memory cache if this is the active collection
-        if !collection_name.is_empty() {
-            let mut cache = self.collections.write().map_err(|e| {
-                tracing::error!(error = ?e, "Collection cache lock poisoned during set_migration");
-                Error::LockPoisoned("collection cache lock poisoned during set_migration".into())
-            })?;
-
-            if let Some(cached) = cache.get(&collection_name) {
-                if cached.schema().version_id == dest_version_id {
-                    cache.insert(collection_name.clone(), Collection::new(dst_col));
-                }
-            }
-        }
-
-        // Rebuild secondary indexes if the destination version is the active collection
-        // and has indexes (matches Go's behavior of reindexing after migration registration)
-        if !collection_name.is_empty() {
-            if let Err(e) = self
-                .maybe_reindex_after_migration(&collection_name, &dest_version_id)
-                .await
-            {}
-        } else {
-        }
-
-        Ok(transform_id)
-    }
-
-    /// Set a migration within an existing transaction context.
-    ///
-    /// This performs the same operations as `set_migration` but uses the provided
-    /// transaction instead of creating a new one. The caller is responsible for
-    /// committing or rolling back the transaction.
-    ///
-    /// This is used for transaction-aware migration configuration via the FFI.
-    pub async fn set_migration_in_txn(
-        &self,
-        txn: &DbTxn<S>,
-        config: LensConfig,
-    ) -> Result<TransformId> {
-        let dest_version_id = config.destination_schema_version_id.clone();
-        let source_version_id = config.source_schema_version_id.clone();
-
-        // Look up source and destination versions, creating placeholders if needed
-        let (source_col, mut dst_col) = {
-            let systemstore = txn.systemstore()?;
-
-            // Look up source version
-            let src_key = CollectionKey::new(&source_version_id);
-            let src_data = systemstore
-                .get(&src_key.bytes())
-                .await
-                .map_err(Error::Storage)?;
-            let source_col: CollectionVersion = match src_data {
-                Some(data) => serde_json::from_slice(&data).map_err(|e| {
-                    Error::Serialization(format!(
-                        "failed to deserialize source schema '{}': {}",
-                        source_version_id, e
-                    ))
-                })?,
-                None => {
-                    // Source doesn't exist — create a placeholder
-                    let mut placeholder = CollectionVersion {
-                        version_id: source_version_id.clone(),
-                        collection_id: ORPHAN_COLLECTION_ID.to_string(),
-                        is_materialized: true,
-                        is_placeholder: true,
-                        ..CollectionVersion::new("", "", "", Vec::new())
-                    };
-                    placeholder.is_active = false;
-                    let data = serde_json::to_vec(&placeholder).map_err(|e| {
-                        Error::Serialization(format!(
-                            "failed to serialize source placeholder '{}': {}",
-                            source_version_id, e
-                        ))
-                    })?;
-                    systemstore
-                        .set(&src_key.bytes(), &data)
-                        .await
-                        .map_err(Error::Storage)?;
-                    placeholder
-                }
-            };
-
-            // Look up destination version
-            let dst_key = CollectionKey::new(&dest_version_id);
-            let dst_data = systemstore
-                .get(&dst_key.bytes())
-                .await
-                .map_err(Error::Storage)?;
-            let dst_col: CollectionVersion = match dst_data {
-                Some(data) => serde_json::from_slice(&data).map_err(|e| {
-                    Error::Serialization(format!(
-                        "failed to deserialize destination schema '{}': {}",
-                        dest_version_id, e
-                    ))
-                })?,
-                None => {
-                    // Destination doesn't exist — create a placeholder
-                    let mut placeholder = CollectionVersion {
-                        name: source_col.name.clone(),
-                        version_id: dest_version_id.clone(),
-                        collection_id: source_col.collection_id.clone(),
-                        is_materialized: true,
-                        is_placeholder: true,
-                        ..CollectionVersion::new("", "", "", Vec::new())
-                    };
-                    placeholder.is_active = false;
-                    let data = serde_json::to_vec(&placeholder).map_err(|e| {
-                        Error::Serialization(format!(
-                            "failed to serialize destination placeholder '{}': {}",
-                            dest_version_id, e
-                        ))
-                    })?;
-                    systemstore
-                        .set(&dst_key.bytes(), &data)
-                        .await
-                        .map_err(Error::Storage)?;
-                    placeholder
-                }
-            };
-
-            (source_col, dst_col)
-        };
-
-        // Validate version adjacency
-        if let Some(ref prev) = dst_col.previous_version {
-            if prev.source_collection_id != source_col.version_id {
-                return Err(Error::InvalidPatch(format!(
-                    "cannot migrate between non-adjacent collection versions. \
-                     Destination '{}' already has previous version '{}', but migration source is '{}'",
-                    dest_version_id, prev.source_collection_id, source_version_id
-                )));
-            }
-        }
-
-        // Register the transform in the lens store
-        let transform_id = self
-            .lens_store
-            .add(config)
-            .await
-            .map_err(|e| Error::Lens(e.to_string()))?;
-
-        // Set the destination's previous_version with source and transform
-        dst_col.previous_version = Some(CollectionSource {
-            source_collection_id: source_col.version_id.clone(),
-            transform: Some(transform_id.to_string()),
-        });
-
-        tracing::debug!(
-            dest_version_id = %dest_version_id,
-            source_version_id = %source_version_id,
-            is_placeholder = dst_col.is_placeholder,
-            transform_id = %transform_id,
-            "set_migration_in_txn: storing destination version with transform"
-        );
-
-        // Save the destination version
-        let collection_name = dst_col.name.clone();
-        let dst_key = CollectionKey::new(&dest_version_id);
-        let dst_data = serde_json::to_vec(&dst_col).map_err(|e| {
-            Error::Serialization(format!(
-                "failed to serialize destination schema '{}': {}",
-                dest_version_id, e
-            ))
-        })?;
-
-        {
-            let systemstore = txn.systemstore()?;
-            systemstore
-                .set(&dst_key.bytes(), &dst_data)
-                .await
-                .map_err(Error::Storage)?;
-
-            // Write CollectionVersionKey entries
-            if !source_col.collection_id.is_empty() {
-                let src_version_key =
-                    CollectionVersionKey::new(&source_col.collection_id, &source_version_id);
-                systemstore
-                    .set(&src_version_key.bytes(), b"1")
-                    .await
-                    .map_err(Error::Storage)?;
-            }
-            if !dst_col.collection_id.is_empty() {
-                let dst_version_key =
-                    CollectionVersionKey::new(&dst_col.collection_id, &dest_version_id);
-                systemstore
-                    .set(&dst_version_key.bytes(), b"1")
-                    .await
-                    .map_err(Error::Storage)?;
-            }
-        }
-
-        // NOTE: We don't commit here - caller is responsible for transaction lifecycle
-
-        // Update in-memory cache if this is the active collection
-        if !collection_name.is_empty() {
-            let mut cache = self.collections.write().map_err(|e| {
-                tracing::error!(error = ?e, "Collection cache lock poisoned during set_migration_in_txn");
-                Error::LockPoisoned("collection cache lock poisoned during set_migration_in_txn".into())
-            })?;
-
-            if let Some(cached) = cache.get(&collection_name) {
-                if cached.schema().version_id == dest_version_id {
-                    cache.insert(collection_name.clone(), Collection::new(dst_col));
-                }
-            }
-        }
-
-        Ok(transform_id)
-    }
-
-    /// Rebuild secondary indexes for a collection after a migration is registered.
-    ///
-    /// If the destination version is the currently active version and the collection
-    /// has indexes, this fetches all documents (applying lens migration), drops
-    /// existing index entries, and rebuilds them with migrated values.
-    async fn maybe_reindex_after_migration(
-        &self,
-        collection_name: &str,
-        dest_version_id: &str,
-    ) -> Result<()> {
-        let collection = match self.get_collection(collection_name)? {
-            Some(c) => c,
-            None => {
-                return Ok(());
-            }
-        };
-
-        // Only reindex if destination is the current active version
-        if collection.version_id() != dest_version_id {
-            return Ok(());
-        }
-
-        // Only reindex if the collection has indexes
-        if collection.get_indexes().is_empty() {
-            return Ok(());
-        }
-
-        let collection_id = collection.collection_id().to_string();
-        let target_version_id = collection.version_id().to_string();
-        let short_id = collection_short_id(&collection_id);
-
-        // Load all versions of this collection to build migration history
-        let read_txn = self.new_txn(true).await?;
-        let systemstore = read_txn.systemstore()?;
-        let versions = get_collections_by_collection_id(&systemstore, &collection_id).await?;
-        let _ = read_txn.discard();
-
-        // Build targeted migration history
-        let history = {
-            let mut full_history: HashMap<String, CollectionHistoryLink> = HashMap::new();
-            for version in &versions {
-                let mut link =
-                    CollectionHistoryLink::new(&version.version_id, &version.collection_id);
-                if let Some(ref prev) = version.previous_version {
-                    link = link.with_previous(&prev.source_collection_id);
-                    if let Some(ref transform_id) = prev.transform {
-                        link = link.with_transform(transform_id);
-                    }
-                }
-                full_history.insert(version.version_id.clone(), link);
-            }
-
-            // Build next links
-            let reverse_links: Vec<(String, String)> = full_history
-                .values()
-                .flat_map(|link| {
-                    link.previous
-                        .iter()
-                        .map(|prev_id| (prev_id.clone(), link.version_id.clone()))
-                        .collect::<Vec<_>>()
-                })
-                .collect();
-            for (parent_id, child_id) in reverse_links {
-                if let Some(parent_link) = full_history.get_mut(&parent_id) {
-                    if !parent_link.next.contains(&child_id) {
-                        parent_link.next.push(child_id);
-                    }
-                }
-            }
-
-            match build_targeted_history(&full_history, &target_version_id) {
-                Some(h) => h,
-                None => return Ok(()),
-            }
-        };
-
-        let has_migrations = history.values().any(|link| link.transform.is_some());
-        if !has_migrations {
-            return Ok(());
-        }
-
-        // Create a write transaction for the reindex
-        let write_txn = self.new_txn(false).await?;
-
-        // Scope the datastore borrow so it's dropped before commit
-        {
-            let datastore = write_txn.datastore()?;
-
-            // Fetch all documents (raw, with their stored schema versions)
-            let raw_docs = collection.get_all_with_datastore(&datastore).await?;
-
-            // Apply lens migration to each document that needs it
-            let mut migrated_docs = Vec::with_capacity(raw_docs.len());
-            for doc in raw_docs {
-                let doc_version = doc
-                    .schema_version_id()
-                    .unwrap_or(&target_version_id)
-                    .to_string();
-
-                if doc_version == target_version_id {
-                    migrated_docs.push(doc);
-                    continue;
-                }
-
-                // Convert to LensDoc
-                if let Ok(map) = doc.to_map() {
-                    let mut lens_doc = LensDoc::new();
-                    for (key, value) in map {
-                        lens_doc.insert(key, value);
-                    }
-
-                    let mut lens =
-                        Lens::new(self.lens_store.clone(), &target_version_id, history.clone());
-
-                    if let Ok(()) = lens.put(&doc_version, lens_doc).await {
-                        if let Some(Ok(migrated_lens_doc)) = lens.next().await {
-                            let mut migrated = document::Document::new();
-                            if let Some(id) = doc.id() {
-                                migrated.set_id(id.clone());
-                            }
-                            for (field_name, value) in migrated_lens_doc {
-                                if field_name != DOC_ID_FIELD {
-                                    // Convert JSON value to native type based on schema field kind
-                                    let native_value = json_to_native_value(
-                                        &value,
-                                        &field_name,
-                                        collection.schema(),
-                                    );
-                                    migrated.set(&field_name, native_value);
-                                }
-                            }
-                            migrated.set_schema_version_id(&target_version_id);
-                            migrated_docs.push(migrated);
-                            continue;
-                        }
-                    }
-                }
-
-                // If migration fails for a doc, keep original
-                migrated_docs.push(doc);
-            }
-
-            // Rebuild indexes: drop all entries, re-index from migrated documents
-            let index_manager = IndexManager::from_collection(short_id, collection.schema())
-                .map_err(|e| Error::Other(format!("failed to create index manager: {}", e)))?;
-
-            for index_desc in collection.get_indexes() {
-                // Drop existing entries
-                if let Some(index) = index_manager.get_index(&index_desc.name) {
-                    index
-                        .remove_all(&mut datastore.clone())
-                        .await
-                        .map_err(Error::Storage)?;
-                }
-
-                // Bulk re-index with migrated documents
-                index_manager
-                    .bulk_index(
-                        &datastore,
-                        &index_desc.name,
-                        &migrated_docs,
-                        collection.schema(),
-                    )
-                    .await?;
-            }
-
-            tracing::debug!(
-                collection = %collection_name,
-                doc_count = migrated_docs.len(),
-                index_count = collection.get_indexes().len(),
-                "Rebuilt indexes after migration"
-            );
-        } // datastore reference dropped here
-
-        write_txn.commit().await?;
-
-        Ok(())
-    }
-
-    /// Check if a migration exists between two schema versions.
-    pub fn has_migration(&self, transform_id: &TransformId) -> bool {
-        self.lens_store.has_transform(transform_id)
-    }
+    // NOTE: Lens migration methods (lens_store, set_migration, set_migration_in_txn,
+    // maybe_reindex_after_migration, has_migration) are now in migration.rs
 
     /// Load all collections from the SystemStore into the in-memory cache.
     ///
@@ -1942,7 +1361,7 @@ impl<S: Store> DB<S> {
     }
 
     /// Get a collection by version ID, searching both cache and KV store.
-    async fn get_collection_by_version_id_full(
+    pub(crate) async fn get_collection_by_version_id_full(
         &self,
         version_id: &str,
     ) -> Result<Option<Collection>> {
@@ -1992,7 +1411,7 @@ impl<S: Store> DB<S> {
     }
 
     /// Check if a collection has any documents in the datastore.
-    async fn collection_has_data(&self, collection_id: &str) -> Result<bool> {
+    pub(crate) async fn collection_has_data(&self, collection_id: &str) -> Result<bool> {
         let txn = self.new_txn(true).await?;
         let has_data = {
             let datastore = txn.datastore()?;
@@ -2005,1734 +1424,6 @@ impl<S: Store> DB<S> {
         };
         let _ = txn.discard();
         Ok(has_data)
-    }
-
-    /// Patch a collection's schema using JSON patch operations.
-    ///
-    /// This creates a new schema version with a new version_id (CID) and links
-    /// it to the previous version via `previous_version`. The old version is
-    /// marked as inactive.
-    ///
-    /// # Arguments
-    ///
-    /// * `collection_name` - The name of the collection to patch
-    /// * `patch` - A JSON patch string (RFC 6902 format)
-    ///
-    /// # Returns
-    ///
-    /// The updated collection version (with new version_id).
-    ///
-    /// # Errors
-    ///
-    /// - `CollectionNotFound` if the collection doesn't exist
-    /// - `InvalidPatch` if the patch is invalid or cannot be applied
-    /// - `Schema` if the resulting schema is invalid
-    pub async fn patch_collection(
-        &self,
-        collection_name: &str,
-        patch: &str,
-    ) -> Result<CollectionVersion> {
-        // Parse the patch early - needed for both collection lookup fallbacks and processing
-        let patch_ops: serde_json::Value =
-            serde_json::from_str(patch).map_err(|e| Error::InvalidPatch(e.to_string()))?;
-
-        // Get the current schema - try by name first, then by version ID (including KV store),
-        // then check for collection-level move/copy targeting a non-existent collection
-        let collection = match self.get_collection(collection_name)? {
-            Some(c) => c,
-            None => {
-                // Try looking up by version ID - search both cache and KV store
-                match self
-                    .get_collection_by_version_id_full(collection_name)
-                    .await?
-                {
-                    Some(c) => c,
-                    None => {
-                        // Collection not found by name or version ID.
-                        // Check if the patch is a collection-level move/copy where the
-                        // "path" targets a non-existent collection (e.g., move /Users → /Books)
-                        return self
-                            .handle_unknown_collection_patch(collection_name, &patch_ops)
-                            .await;
-                    }
-                }
-            }
-        };
-
-        let old_schema = collection.schema().clone();
-        let actual_name = old_schema.name.clone();
-        let old_version_id = old_schema.version_id.clone();
-        let collection_id = old_schema.collection_id.clone();
-
-        // Collect known collection names for Kind validation
-        let known_collection_names: Vec<String> = self
-            .list_collections()
-            .unwrap_or_default()
-            .into_iter()
-            .collect();
-
-        // Apply the patch to the schema JSON
-        let mut schema_json = serde_json::to_value(&old_schema).map_err(|e| {
-            Error::Serialization(format!("failed to serialize schema to JSON: {}", e))
-        })?;
-
-        // Normalize JSON to match Go's serialization format before applying patches.
-        // Go always serializes struct fields (null for nil pointers, [] for nil slices),
-        // but Rust's skip_serializing_if omits them. Patches targeting these paths
-        // need the keys to exist for replace/remove operations to work correctly,
-        // and for validators to run instead of json_pointer errors.
-        // Note: EncryptedIndexes is NOT pre-populated because Go doesn't expose
-        // it in the JSON representation - patches targeting it should fail.
-        if let serde_json::Value::Object(ref mut map) = schema_json {
-            for key in &["Indexes", "VectorEmbeddings"] {
-                map.entry(key.to_string())
-                    .or_insert(serde_json::Value::Array(vec![]));
-            }
-            for key in &["CollectionSet", "Query", "PreviousVersion", "Policy"] {
-                map.entry(key.to_string())
-                    .or_insert(serde_json::Value::Null);
-            }
-        }
-
-        // Apply JSON patch operations
-        // Go DefraDB embeds collection name in patch paths: /CollectionName/Fields/-
-        // We need to strip the collection name prefix to get paths relative to schema.
-        // Patches may use the collection name, actual name, or version ID as prefix.
-        // Build a list of all recognized prefixes to try stripping.
-        let mut strip_prefixes: Vec<String> = vec![format!("/{}/", collection_name)];
-        if actual_name != collection_name {
-            strip_prefixes.push(format!("/{}/", actual_name));
-        }
-        if old_version_id != collection_name && old_version_id != actual_name {
-            strip_prefixes.push(format!("/{}/", old_version_id));
-        }
-
-        // Track whether the patch deactivates this collection or explicitly changes IsActive.
-        // These require in-place updates rather than new version creation.
-        let mut is_deactivation = false;
-        let mut is_active_explicitly_set = false;
-
-        if let serde_json::Value::Array(ops) = patch_ops {
-            for op in ops {
-                let operation = op.get("op").and_then(|v| v.as_str());
-                let raw_path = op.get("path").and_then(|v| v.as_str());
-                let value = op.get("value");
-
-                // Strip collection name/version prefix from path if present (Go compatibility)
-                let stripped_path =
-                    raw_path.map(|p| Self::strip_collection_prefix(p, &strip_prefixes));
-
-                // Extract field name from path before substitution (for name mismatch validation)
-                let field_name_from_path = stripped_path
-                    .as_deref()
-                    .and_then(|p| Self::extract_field_name_from_path(p));
-
-                // Go compatibility: substitute field names for indices in /Fields/<name> paths
-                let path =
-                    stripped_path.map(|p| Self::substitute_field_name_in_path(&p, &schema_json));
-
-                match (operation, path.as_deref()) {
-                    (Some("replace"), Some(path)) | (Some("add"), Some(path)) => {
-                        let mut value = value
-                            .ok_or_else(|| {
-                                Error::InvalidPatch(format!(
-                                    "missing 'value' for operation at {}",
-                                    path
-                                ))
-                            })?
-                            .clone();
-
-                        // Go compatibility: root-level add/replace is "adding collections"
-                        if path == "/" {
-                            let name = value
-                                .as_object()
-                                .and_then(|m| m.get("Name"))
-                                .and_then(|n| n.as_str())
-                                .unwrap_or(&actual_name);
-                            return Err(Error::InvalidPatch(format!(
-                                "adding collections via patch is not supported. Name: {}",
-                                name
-                            )));
-                        }
-
-                        // Track explicit IsActive changes for in-place update handling
-                        if path == "/IsActive" {
-                            is_active_explicitly_set = true;
-                        }
-
-                        // Go compatibility: validate VersionID replacement
-                        if path.ends_with("/VersionID") {
-                            let version_id_str = value.as_str().unwrap_or("");
-                            if version_id_str.is_empty() {
-                                return Err(Error::InvalidPatch(
-                                    "collection ID cannot be empty".to_string(),
-                                ));
-                            }
-                            // Validate CID format
-                            if cid::Cid::try_from(version_id_str).is_err() {
-                                return Err(Error::InvalidPatch(format!(
-                                    "invalid cid: selected encoding not supported. VersionID: {}",
-                                    version_id_str
-                                )));
-                            }
-                            // Check if this CID exists as a known collection version
-                            let all_versions =
-                                self.get_all_collection_versions().await.unwrap_or_default();
-                            let is_known =
-                                all_versions.iter().any(|c| c.version_id == version_id_str);
-                            if !is_known {
-                                return Err(Error::InvalidPatch(
-                                    "unknown CID, collection ids cannot be manually defined"
-                                        .to_string(),
-                                ));
-                            }
-                            // Known CIDs proceed - sources/ownership validation
-                            // is handled by definition_validation post-patch.
-                        }
-
-                        // Go compatibility: validate Kind value when replacing a field's Kind directly
-                        // e.g., replace /Fields/2/Kind with "NotAValidKind"
-                        if path.ends_with("/Kind") && path.contains("/Fields/") {
-                            // Extract field name from the path for error message
-                            // Path is like /Fields/2/Kind, we need to get the name of field at index 2
-                            let field_index_str = path
-                                .trim_start_matches("/Fields/")
-                                .split('/')
-                                .next()
-                                .unwrap_or("");
-                            let field_name = if let Ok(idx) = field_index_str.parse::<usize>() {
-                                schema_json
-                                    .get("Fields")
-                                    .and_then(|f| f.as_array())
-                                    .and_then(|arr| arr.get(idx))
-                                    .and_then(|f| f.get("Name"))
-                                    .and_then(|n| n.as_str())
-                                    .unwrap_or("")
-                            } else {
-                                field_index_str
-                            };
-                            Self::validate_patch_field_kind(
-                                &value,
-                                field_name,
-                                &known_collection_names,
-                            )?;
-                        }
-
-                        // Go compatibility: validate and auto-generate FieldID when adding new fields
-                        // If path ends with /Fields/- or /Fields/<n> and value has Name but no FieldID
-                        if path.contains("/Fields/") {
-                            if let serde_json::Value::Object(ref mut map) = value {
-                                // Validate field name matches path index name (Go compatibility)
-                                if let Some(ref path_name) = field_name_from_path {
-                                    if let Some(value_name) =
-                                        map.get("Name").and_then(|n| n.as_str())
-                                    {
-                                        if !value_name.is_empty()
-                                            && value_name != path_name.as_str()
-                                        {
-                                            return Err(Error::InvalidPatch(format!(
-                                                "the index used does not match the given name. index: {}, name: {}",
-                                                path_name, value_name
-                                            )));
-                                        }
-                                    }
-                                    // If value doesn't have Name, set it from the path
-                                    if !map.contains_key("Name") {
-                                        map.insert(
-                                            "Name".to_string(),
-                                            serde_json::Value::String(path_name.clone()),
-                                        );
-                                    }
-                                }
-
-                                // Validate Kind value for new fields
-                                if let Some(kind_val) = map.get("Kind") {
-                                    Self::validate_patch_field_kind(
-                                        kind_val,
-                                        map.get("Name").and_then(|n| n.as_str()).unwrap_or(""),
-                                        &known_collection_names,
-                                    )?;
-                                }
-                                if map.contains_key("Name") && !map.contains_key("FieldID") {
-                                    // Find max existing FieldID to avoid collisions with gaps
-                                    let max_field_id = schema_json
-                                        .get("Fields")
-                                        .and_then(|f| f.as_array())
-                                        .map(|arr| {
-                                            arr.iter()
-                                                .filter_map(|f| f.get("FieldID"))
-                                                .filter_map(|id| {
-                                                    id.as_str()
-                                                        .and_then(|s| s.parse::<u64>().ok())
-                                                        .or_else(|| id.as_u64())
-                                                })
-                                                .max()
-                                                .unwrap_or(0)
-                                        })
-                                        .unwrap_or(0);
-                                    let field_id = (max_field_id + 1).to_string();
-                                    map.insert("FieldID".to_string(), field_id.into());
-                                }
-                            }
-                        }
-
-                        // Go compatibility: For top-level array fields that don't exist (like
-                        // EncryptedIndexes which isn't exposed in Go's JSON), produce Go-compatible
-                        // error messages.
-                        let is_top_level_path = path.starts_with('/') && !path[1..].contains('/');
-                        if is_top_level_path {
-                            let key = &path[1..];
-                            let key_exists = schema_json
-                                .as_object()
-                                .map(|m| m.contains_key(key))
-                                .unwrap_or(false);
-                            if !key_exists {
-                                // For add with array value, Go produces unmarshal error
-                                if operation == Some("add") && value.is_array() {
-                                    return Err(Error::InvalidPatch(
-                                        "cannot unmarshal array into Go value".to_string(),
-                                    ));
-                                }
-                                // The "-" key is the JSON Patch append operator. When used
-                                // on an object (not array), Go applies it then fails during
-                                // unmarshalling because "-" is not a valid struct field.
-                                if operation == Some("add") && key == "-" {
-                                    return Err(Error::InvalidPatch(
-                                        "json: unknown field \"-\"".to_string(),
-                                    ));
-                                }
-                                // For replace on non-existent key, Go produces "doc is missing key"
-                                if operation == Some("replace") {
-                                    return Err(Error::InvalidPatch(
-                                        "doc is missing key".to_string(),
-                                    ));
-                                }
-                            }
-                        }
-
-                        if let Err(e) = Self::json_pointer_set(&mut schema_json, path, value) {
-                            match &e {
-                                Error::InvalidPatch(msg)
-                                    if msg.starts_with("path not found:")
-                                        || msg.starts_with("cannot set value")
-                                        || msg.starts_with("cannot navigate") =>
-                                {
-                                    return Err(Error::InvalidPatch(
-                                        "add operation does not apply: doc is missing path"
-                                            .to_string(),
-                                    ));
-                                }
-                                _ => return Err(e),
-                            }
-                        }
-                    }
-                    (Some("remove"), Some(path)) => {
-                        if path == "/" {
-                            // Root-level remove = deactivate collection
-                            is_deactivation = true;
-                        } else {
-                            // Go compatibility: For top-level keys that don't exist (like
-                            // EncryptedIndexes), produce Go-compatible error message.
-                            let is_top_level_path =
-                                path.starts_with('/') && !path[1..].contains('/');
-                            if is_top_level_path {
-                                let key = &path[1..];
-                                let key_exists = schema_json
-                                    .as_object()
-                                    .map(|m| m.contains_key(key))
-                                    .unwrap_or(false);
-                                if !key_exists {
-                                    return Err(Error::InvalidPatch(
-                                        "unable to remove nonexistent key".to_string(),
-                                    ));
-                                }
-                            }
-                            Self::json_pointer_remove(&mut schema_json, path)?;
-                        }
-                    }
-                    (Some("test"), Some(path)) => {
-                        // RFC 6902 "test" operation: verify value at path equals expected
-                        let expected_value = value
-                            .ok_or_else(|| {
-                                Error::InvalidPatch(format!(
-                                    "missing 'value' for test operation at {}",
-                                    path
-                                ))
-                            })?
-                            .clone();
-
-                        // Get the actual value at the path
-                        let actual_value = Self::json_pointer_get(&schema_json, path);
-
-                        // Compare: if path doesn't exist or values don't match, test fails
-                        match actual_value {
-                            Some(actual) if actual == expected_value => {
-                                // Test passes - continue to next operation
-                            }
-                            _ => {
-                                // Test fails - return error in Go-compatible format
-                                // Include original path for context
-                                let original_path = raw_path.unwrap_or(path);
-                                return Err(Error::InvalidPatch(format!(
-                                    "testing value {} failed: test failed",
-                                    original_path
-                                )));
-                            }
-                        }
-                    }
-                    (Some("copy"), Some(path)) => {
-                        // RFC 6902 "copy" operation: copy value from "from" to "path"
-                        let from_path =
-                            op.get("from").and_then(|v| v.as_str()).ok_or_else(|| {
-                                Error::InvalidPatch(format!(
-                                    "missing 'from' for copy operation at {}",
-                                    path
-                                ))
-                            })?;
-
-                        // Go compatibility: copying collection-level is not supported
-                        // This includes copying to root "/" or to paths that would create new collections
-                        // Detect by checking if path doesn't contain /Fields (field-level operations)
-                        if path == "/"
-                            || (!path.contains("/Fields")
-                                && !path.contains("/Name")
-                                && !path.contains("/IsActive"))
-                        {
-                            // Extract the target name from the raw path for the error message
-                            let target_name = raw_path
-                                .and_then(|p| {
-                                    let p = p.trim_start_matches('/');
-                                    p.split('/').next()
-                                })
-                                .unwrap_or("Unknown");
-                            return Err(Error::InvalidPatch(format!(
-                                "adding collections via patch is not supported. Name: {}",
-                                target_name
-                            )));
-                        }
-
-                        // Substitute field names in from path too
-                        let from_path =
-                            Self::substitute_field_name_in_path(from_path, &schema_json);
-                        // Strip collection prefix from "from" path if present
-                        let from_path = Self::strip_collection_prefix(&from_path, &strip_prefixes);
-
-                        // Get the value to copy. First try current schema, then
-                        // cross-collection: Go applies patches against a global dict
-                        // of all collections, so "from" can reference other collections.
-                        let value_to_copy = Self::json_pointer_get(&schema_json, &from_path)
-                            .or_else(|| {
-                                let trimmed = from_path.trim_start_matches('/');
-                                let first_slash = trimmed.find('/');
-                                if let Some(pos) = first_slash {
-                                    let other_name = &trimmed[..pos];
-                                    let rest = &trimmed[pos..];
-                                    if let Ok(Some(other_col)) = self.get_collection(other_name) {
-                                        let other_schema = other_col.schema();
-                                        if let Ok(other_json) = serde_json::to_value(other_schema) {
-                                            return Self::json_pointer_get(&other_json, rest);
-                                        }
-                                    }
-                                }
-                                None
-                            })
-                            .ok_or_else(|| {
-                                Error::InvalidPatch(format!("path not found: {}", from_path))
-                            })?;
-
-                        // When copying a field, clear the FieldID so it becomes a "new" field.
-                        // Go DefraDB generates new FieldIDs for copied fields rather than
-                        // treating them as mutations of the original.
-                        let value_to_copy =
-                            if path.contains("/Fields/") && value_to_copy.is_object() {
-                                let mut v = value_to_copy;
-                                if let Some(obj) = v.as_object_mut() {
-                                    obj.remove("FieldID");
-                                }
-                                v
-                            } else {
-                                value_to_copy
-                            };
-
-                        // Set at destination
-                        Self::json_pointer_set(&mut schema_json, path, value_to_copy)?;
-                    }
-                    (Some("move"), Some(path)) => {
-                        // RFC 6902 "move" operation: move value from "from" to "path"
-                        let from_path =
-                            op.get("from").and_then(|v| v.as_str()).ok_or_else(|| {
-                                Error::InvalidPatch(format!(
-                                    "missing 'from' for move operation at {}",
-                                    path
-                                ))
-                            })?;
-
-                        // Go compatibility: moving at collection-level is a no-op
-                        // This includes moving to root "/" or paths that would move entire collections
-                        // Detect by checking if path doesn't contain /Fields (field-level operations)
-                        if path == "/"
-                            || (!path.contains("/Fields")
-                                && !path.contains("/Name")
-                                && !path.contains("/IsActive"))
-                        {
-                            // Skip this operation - collection-level moves are no-ops
-                            continue;
-                        }
-
-                        // Substitute field names in from path too
-                        let from_path =
-                            Self::substitute_field_name_in_path(from_path, &schema_json);
-                        // Strip collection prefix from "from" path if present
-                        let from_path = Self::strip_collection_prefix(&from_path, &strip_prefixes);
-
-                        // Get the value to move
-                        let value_to_move = Self::json_pointer_get(&schema_json, &from_path)
-                            .ok_or_else(|| {
-                                Error::InvalidPatch(format!("path not found: {}", from_path))
-                            })?;
-
-                        // Remove from source first
-                        Self::json_pointer_remove(&mut schema_json, &from_path)?;
-
-                        // Set at destination
-                        Self::json_pointer_set(&mut schema_json, path, value_to_move)?;
-                    }
-                    _ => {
-                        return Err(Error::InvalidPatch(format!(
-                            "unsupported or invalid patch operation: {:?}",
-                            op
-                        )));
-                    }
-                }
-            }
-        } else {
-            return Err(Error::InvalidPatch(
-                "patch must be an array of operations".to_string(),
-            ));
-        }
-
-        // Go compatibility: auto-generate FieldID for any fields missing one
-        // This handles cases where FieldID is removed (e.g., after copy operation)
-        if let Some(fields) = schema_json.get_mut("Fields").and_then(|f| f.as_array_mut()) {
-            // Find max existing FieldID
-            let max_field_id: u64 = fields
-                .iter()
-                .filter_map(|f| f.get("FieldID"))
-                .filter_map(|id| {
-                    id.as_str()
-                        .and_then(|s| s.parse::<u64>().ok())
-                        .or_else(|| id.as_u64())
-                })
-                .max()
-                .unwrap_or(0);
-
-            let mut next_id = max_field_id + 1;
-            for field in fields.iter_mut() {
-                if let serde_json::Value::Object(ref mut map) = field {
-                    if !map.contains_key("FieldID")
-                        || map.get("FieldID") == Some(&serde_json::Value::Null)
-                        || map.get("FieldID").and_then(|v| v.as_str()) == Some("")
-                    {
-                        map.insert("FieldID".to_string(), next_id.to_string().into());
-                        next_id += 1;
-                    }
-                }
-            }
-        }
-
-        // Go compatibility: check for empty collection name before deserialization
-        let name_value = schema_json.get("Name");
-        match name_value {
-            None | Some(serde_json::Value::Null) => {
-                return Err(Error::InvalidPatch(
-                    "collection name can't be empty".to_string(),
-                ));
-            }
-            Some(serde_json::Value::String(s)) if s.is_empty() => {
-                return Err(Error::InvalidPatch(
-                    "collection name can't be empty".to_string(),
-                ));
-            }
-            _ => {}
-        }
-
-        // Field movement and duplicate checks are handled by the definition validators
-        // (validate_field_not_moved, validate_field_not_duplicated) which run post-deserialization.
-        // This matches Go's approach of collecting ALL validation errors at once.
-
-        // Deserialize back to CollectionVersion
-        let mut new_schema: CollectionVersion = serde_json::from_value(schema_json)
-            .map_err(|e| Error::InvalidPatch(format!("invalid resulting schema: {}", e)))?;
-
-        // Go compatibility: check for removed/empty required fields after deserialization.
-        // Go's JSON unmarshaling uses zero values for missing fields; our serde defaults
-        // replicate this. Now check for invalid empty values that indicate patch corruption.
-        if new_schema.version_id.is_empty() {
-            return Err(Error::InvalidPatch(
-                "invalid cid: cid too short. VersionID: ".to_string(),
-            ));
-        }
-
-        // Go compatibility: validate field Kind and Name after patching.
-        // Existing fields (matched by FieldID) that were modified are "mutations"
-        // which is not supported. New fields with Kind:0 or empty Name are separate errors.
-        {
-            let old_field_ids: std::collections::HashSet<&str> =
-                old_schema.fields.iter().map(|f| f.id.as_str()).collect();
-            let old_field_map: std::collections::HashMap<&str, &schema::FieldDescription> =
-                old_schema
-                    .fields
-                    .iter()
-                    .map(|f| (f.id.as_str(), f))
-                    .collect();
-            let mut field_errors = Vec::new();
-
-            for field in &new_schema.fields {
-                let is_old_field =
-                    !field.id.is_empty() && old_field_ids.contains(field.id.as_str());
-
-                if is_old_field {
-                    // Check if an existing field was mutated (Kind changed, Name removed, etc.)
-                    if let Some(old_field) = old_field_map.get(field.id.as_str()) {
-                        let kind_changed = field.kind != old_field.kind;
-                        let name_empty = field.name.is_empty();
-                        if kind_changed || name_empty {
-                            field_errors.push(
-                                "mutating an existing field is not supported. ProposedName: "
-                                    .to_string(),
-                            );
-                        }
-                    }
-                } else {
-                    // Check if this is a relational ID field (_<name>ID pattern)
-                    // A field is a relational ID if:
-                    // 1. Its name matches _<base>ID where <base> is another field name
-                    // 2. That base field has a relation-like Kind (Named, Relation, or SelfRef)
-                    // Note: we use is_relation() which checks the Kind variant directly
-                    let is_relational_id = field
-                        .name
-                        .strip_prefix('_')
-                        .and_then(|s| s.strip_suffix("ID"))
-                        .map(|base_name| {
-                            new_schema.fields.iter().any(|f| {
-                                f.name == base_name
-                                    && matches!(
-                                        f.kind,
-                                        schema::FieldKind::Named { .. }
-                                            | schema::FieldKind::Relation { .. }
-                                            | schema::FieldKind::SelfRef { .. }
-                                    )
-                            })
-                        })
-                        .unwrap_or(false);
-
-                    if is_relational_id {
-                        // Relational ID fields must have Kind = DocID (1)
-                        if !matches!(
-                            field.kind,
-                            schema::FieldKind::Scalar(schema::ScalarKind::DocID)
-                        ) {
-                            let kind_display = match &field.kind {
-                                schema::FieldKind::Scalar(k) => match k {
-                                    schema::ScalarKind::None => "0".to_string(),
-                                    schema::ScalarKind::DocID => "ID".to_string(),
-                                    schema::ScalarKind::Bool => "Boolean".to_string(),
-                                    schema::ScalarKind::Int => "Integer".to_string(),
-                                    schema::ScalarKind::Float64 => "Float".to_string(),
-                                    schema::ScalarKind::Float32 => "Float32".to_string(),
-                                    schema::ScalarKind::DateTime => "DateTime".to_string(),
-                                    schema::ScalarKind::String => "String".to_string(),
-                                    schema::ScalarKind::Blob => "Blob".to_string(),
-                                    schema::ScalarKind::Json => "JSON".to_string(),
-                                },
-                                other => format!("{:?}", other),
-                            };
-                            field_errors.push(format!(
-                                "relational id field of invalid kind. Field: {}, Expected: ID, Actual: {}",
-                                field.name, kind_display
-                            ));
-                        }
-                    } else {
-                        // Standard new field validations
-                        if matches!(
-                            field.kind,
-                            schema::FieldKind::Scalar(schema::ScalarKind::None)
-                        ) && field.name != "_docID"
-                        {
-                            field_errors.push(format!(
-                                "no type found for given name. Type: {}",
-                                schema::ScalarKind::None as u8
-                            ));
-                        }
-                        if field.name.is_empty() {
-                            field_errors.push(
-                                "Names must match /^[_a-zA-Z][_a-zA-Z0-9]*$/ but \"\" does not."
-                                    .to_string(),
-                            );
-                        }
-                    }
-                }
-            }
-
-            if !field_errors.is_empty() {
-                return Err(Error::InvalidPatch(field_errors.join("\n")));
-            }
-        }
-
-        // Go compatibility: auto-generate _fieldID for foreign object fields added via patch.
-        // This matches Go's collection_define.go behavior for fields with Kind.IsObject() && !Kind.IsArray().
-        {
-            let max_field_id: u64 = new_schema
-                .fields
-                .iter()
-                .filter_map(|f| f.id.parse::<u64>().ok())
-                .max()
-                .unwrap_or(0);
-            let mut next_id = max_field_id + 1;
-            new_schema
-                .add_relation_id_fields(|| {
-                    let id = next_id.to_string();
-                    next_id += 1;
-                    id
-                })
-                .map_err(|e| {
-                    Error::InvalidPatch(format!("failed to add relation id fields: {}", e))
-                })?;
-        }
-
-        // Go compatibility: validate self-referential relation field completeness.
-        // For self-references (where field.Kind references the same collection), the schema
-        // must have another field with the same RelationName (the other side of the relation).
-        // This only applies to self-references; cross-collection relations are validated
-        // at a later stage after all patches have been applied.
-        {
-            let mut rel_errors = Vec::new();
-            for field in &new_schema.fields {
-                if !field.kind.is_relation() {
-                    continue;
-                }
-                let relation_name = match &field.relation_name {
-                    Some(rn) => rn.clone(),
-                    None => continue,
-                };
-
-                // Get the referenced collection name from the Kind
-                let ref_name = match &field.kind {
-                    schema::FieldKind::Named { name, .. } => name.clone(),
-                    schema::FieldKind::Relation { collection_id, .. } => collection_id.clone(),
-                    schema::FieldKind::SelfRef { .. } => new_schema.name.clone(),
-                    _ => continue,
-                };
-
-                // Only check self-references (where the Kind references this same collection)
-                let is_self_ref = ref_name == new_schema.name || ref_name == actual_name;
-                if !is_self_ref {
-                    continue;
-                }
-
-                // Check if there's another field with the same relation name
-                let has_counterpart = new_schema.fields.iter().any(|f| {
-                    f.name != field.name
-                        && f.relation_name.as_deref() == Some(relation_name.as_str())
-                });
-
-                if !has_counterpart {
-                    rel_errors.push(format!(
-                        "relation missing field. Object: {}, RelationName: {}",
-                        ref_name, relation_name
-                    ));
-                }
-            }
-
-            if !rel_errors.is_empty() {
-                return Err(Error::InvalidPatch(rel_errors.join("\n")));
-            }
-        }
-
-        // Handle in-place updates (deactivation, IsActive-only, or PreviousVersion/Transform-only).
-        // These don't create a new schema version - they update the existing one.
-        let is_isactive_only_change = is_active_explicitly_set
-            && new_schema.fields == old_schema.fields
-            && new_schema.name == old_schema.name;
-
-        // Check if only PreviousVersion/Transform changed (lens migration linking).
-        // This is an in-place update that adds a migration transform to an existing version.
-        let is_transform_only_change = !is_deactivation
-            && !is_active_explicitly_set
-            && new_schema.fields == old_schema.fields
-            && new_schema.name == old_schema.name
-            && new_schema.is_active == old_schema.is_active
-            && new_schema.previous_version != old_schema.previous_version;
-
-        if is_deactivation || is_isactive_only_change || is_transform_only_change {
-            if is_deactivation {
-                new_schema.is_active = false;
-            }
-            // Keep original version_id
-            new_schema.version_id = old_version_id.clone();
-            // For IsActive-only or deactivation, restore original previous_version.
-            // For Transform-only changes, keep the new previous_version (contains the transform).
-            if !is_transform_only_change {
-                new_schema.previous_version = old_schema.previous_version.clone();
-            }
-
-            // Validate: can't remove a version that is a dependency of another version
-            // This check runs always for deactivation (even if already inactive),
-            // matching Go's validateCollectionDoesNotHaveHigherVersion
-            if is_deactivation {
-                let all_versions = self.get_all_collection_versions().await?;
-                for other in &all_versions {
-                    if let Some(ref prev) = other.previous_version {
-                        if prev.source_collection_id == old_version_id {
-                            return Err(Error::InvalidPatch(
-                                "cannot delete a version that is used by a newer version, first delete the new version".to_string(),
-                            ));
-                        }
-                    }
-                }
-            }
-
-            // Validate: can't remove a collection that has documents (only on active→inactive)
-            if !new_schema.is_active && old_schema.is_active {
-                let has_data = self.collection_has_data(&collection_id).await?;
-                if has_data {
-                    return Err(Error::InvalidPatch(
-                        "cannot delete a collection that has documents, first delete the documents and then delete the version".to_string(),
-                    ));
-                }
-            }
-
-            // Run cross-collection validators to catch issues like multiple active versions
-            let all_existing = self.get_all_collection_versions().await?;
-            let new_collections: Vec<CollectionVersion> = all_existing
-                .iter()
-                .filter(|c| c.version_id != old_version_id)
-                .cloned()
-                .chain(std::iter::once(new_schema.clone()))
-                .collect();
-            crate::definition_validation::validate_collection_changes(
-                &all_existing,
-                &new_collections,
-            )
-            .map_err(Error::InvalidPatch)?;
-
-            // Store the updated version
-            let txn = self.new_txn(false).await?;
-            {
-                let systemstore = txn.systemstore()?;
-                let key = CollectionKey::new(&old_version_id);
-                let data = serde_json::to_vec(&new_schema).map_err(|e| {
-                    Error::Serialization(format!(
-                        "failed to serialize updated schema version '{}': {}",
-                        old_version_id, e
-                    ))
-                })?;
-                systemstore
-                    .set(&key.bytes(), &data)
-                    .await
-                    .map_err(Error::Storage)?;
-
-                // Update name pointer based on activation state
-                let name_key = CollectionNameKey::new(&actual_name);
-                if new_schema.is_active {
-                    systemstore
-                        .set(&name_key.bytes(), old_version_id.as_bytes())
-                        .await
-                        .map_err(Error::Storage)?;
-                } else {
-                    systemstore
-                        .delete(&name_key.bytes())
-                        .await
-                        .map_err(Error::Storage)?;
-                }
-            }
-            txn.commit().await?;
-
-            // Update cache
-            let mut cache = self.collections.write().map_err(|e| {
-                tracing::error!(error = ?e, "Collection cache lock poisoned during in-place update");
-                Error::CacheUpdateFailedAfterCommit(actual_name.clone())
-            })?;
-            if new_schema.is_active {
-                cache.insert(actual_name.clone(), Collection::new(new_schema.clone()));
-            } else {
-                cache.remove(&actual_name);
-            }
-
-            tracing::info!(
-                collection = %actual_name,
-                version = %old_version_id,
-                is_active = new_schema.is_active,
-                "Updated collection version in place"
-            );
-
-            return Ok(new_schema);
-        }
-
-        // --- Normal path: create a new schema version ---
-
-        // Go compatibility: default new fields with CType::None to CType::LwwRegister.
-        // Go's patchCollection does this in collection_define.go for new fields that
-        // don't have an explicit CRDT type. This must happen before CID generation.
-        {
-            let old_field_names: std::collections::HashSet<&str> =
-                old_schema.fields.iter().map(|f| f.name.as_str()).collect();
-            for field in &mut new_schema.fields {
-                if !old_field_names.contains(field.name.as_str())
-                    && field.crdt_type == schema::CType::None
-                {
-                    field.crdt_type = schema::CType::LwwRegister;
-                }
-            }
-        }
-
-        // Run Go-compatible cross-collection validators (before schema validate() which
-        // uses different error messages). These validators cover duplicate fields,
-        // CRDT/kind compatibility, and all Go-specific patch constraints.
-        let all_existing = self.get_all_collection_versions().await?;
-        let new_collections: Vec<CollectionVersion> = all_existing
-            .iter()
-            .filter(|c| c.version_id != old_version_id)
-            .cloned()
-            .chain(std::iter::once(new_schema.clone()))
-            .collect();
-        crate::definition_validation::validate_collection_changes(&all_existing, &new_collections)
-            .map_err(Error::InvalidPatch)?;
-
-        // Also run schema-level validation for checks not covered by definition validators
-        // (e.g., relation field requires relation_name, policy format validation)
-        new_schema.validate()?;
-
-        // Auto-create unique indexes for one-to-one relations added via patch.
-        // This runs AFTER validation (which rejects index mutations on existing schemas)
-        // but BEFORE CID generation (since indexes are part of the schema content).
-        {
-            // Go uses sequential IDs starting from the next available for this collection
-            let schema_max_index_id = new_schema
-                .indexes
-                .iter()
-                .map(|idx| idx.id)
-                .max()
-                .unwrap_or(0);
-            let mut next_index_id = schema_max_index_id;
-
-            let mut indexes_to_add = Vec::new();
-            for field in &new_schema.fields {
-                if !field.kind.is_relation() || field.kind.is_array() {
-                    continue;
-                }
-                let rel_name = match field.relation_name.as_ref() {
-                    Some(n) => n,
-                    None => continue,
-                };
-                let other_col_id = match field.kind.relation_collection_id() {
-                    Some(id) => id,
-                    None => continue,
-                };
-                // Look up the other collection (may be the same collection for self-ref)
-                let other_col = all_existing.iter().find(|c| {
-                    (c.name == other_col_id || c.collection_id == other_col_id) && c.is_active
-                });
-                if let Some(other_col) = other_col {
-                    let other_field =
-                        other_col.field_by_relation(rel_name, &new_schema.name, &field.name);
-                    let other_is_array = other_field.map(|f| f.kind.is_array()).unwrap_or(false);
-                    // One-to-one: other side exists and is non-array, this field is primary
-                    if !other_is_array && field.is_primary {
-                        match new_schema.ensure_one_to_one_unique_index(&field.name, &mut || {
-                            next_index_id += 1;
-                            next_index_id
-                        }) {
-                            Ok(Some(index)) => indexes_to_add.push(index),
-                            Ok(None) => {} // existing unique index is fine
-                            Err(e) => return Err(Error::InvalidPatch(e.to_string())),
-                        }
-                    }
-                }
-            }
-            for index in indexes_to_add {
-                new_schema.indexes.push(index);
-            }
-        }
-
-        // Compute version depth by following the previous_version chain from old_schema to root.
-        // This ensures branching schemas get the correct priority (based on chain depth,
-        // not total version count). Both branches from the same parent get the same priority.
-        let version_depth = {
-            let versions_map: std::collections::HashMap<&str, &CollectionVersion> = all_existing
-                .iter()
-                .map(|v| (v.version_id.as_str(), v))
-                .collect();
-            let mut depth = 0u64;
-            let mut current_id = old_schema.version_id.as_str();
-            loop {
-                match versions_map.get(current_id) {
-                    Some(v) => match &v.previous_version {
-                        Some(prev) => {
-                            depth += 1;
-                            current_id = prev.source_collection_id.as_str();
-                        }
-                        None => break,
-                    },
-                    None => break,
-                }
-            }
-            // depth = chain length from old_schema to root
-            // version_depth = depth + 1 (the old version's "height" in the DAG)
-            depth + 1
-        };
-
-        // Generate new version_id from schema content with proper priorities
-        let new_version_id =
-            Self::generate_patch_version_id(&mut new_schema, &old_schema, version_depth);
-
-        // Update new schema with version info
-        new_schema.version_id = new_version_id.clone();
-
-        // Check if a placeholder version exists with this ID (from pre-registered migration).
-        // When set_migration is called before patch_collection, it creates a placeholder
-        // with previous_version.transform set. We need to copy that transform to preserve
-        // the migration link.
-        let placeholder_transform = {
-            let read_txn = self.new_txn(true).await?;
-            let systemstore = read_txn.systemstore()?;
-            let placeholder_key = CollectionKey::new(&new_version_id);
-            match systemstore
-                .get(&placeholder_key.bytes())
-                .await
-                .map_err(Error::Storage)?
-            {
-                Some(data) => {
-                    let placeholder: CollectionVersion =
-                        serde_json::from_slice(&data).map_err(|e| {
-                            Error::Serialization(format!(
-                                "failed to deserialize placeholder version: {}",
-                                e
-                            ))
-                        })?;
-                    tracing::debug!(
-                        new_version_id = %new_version_id,
-                        is_placeholder = placeholder.is_placeholder,
-                        has_previous_version = placeholder.previous_version.is_some(),
-                        transform = ?placeholder.previous_version.as_ref().and_then(|pv| pv.transform.as_ref()),
-                        "patch_collection: found existing version"
-                    );
-                    if placeholder.is_placeholder {
-                        // Found a placeholder - extract its transform
-                        placeholder.previous_version.and_then(|pv| pv.transform)
-                    } else {
-                        None
-                    }
-                }
-                None => None,
-            }
-        };
-
-        // Use placeholder transform if available, otherwise None
-        new_schema.previous_version = Some(CollectionSource {
-            source_collection_id: old_version_id.clone(),
-            transform: placeholder_transform.clone(),
-        });
-
-        if placeholder_transform.is_some() {
-            tracing::debug!(
-                new_version = %new_version_id,
-                transform_id = ?placeholder_transform,
-                "Linked pre-registered migration from placeholder to new schema version"
-            );
-        }
-
-        // Also check for pending migrations targeting this new version (in-memory fallback)
-        {
-            let pending = self.pending_migrations.read().map_err(|e| {
-                tracing::error!(error = ?e, "Pending migrations lock poisoned");
-                Error::LockPoisoned(
-                    "pending migrations lock poisoned during patch_collection".into(),
-                )
-            })?;
-            if let Some((_source_id, transform_id)) = pending.get(&new_version_id) {
-                if let Some(ref mut prev) = new_schema.previous_version {
-                    // Only override if we didn't already get a transform from the placeholder
-                    if prev.transform.is_none() {
-                        prev.transform = Some(transform_id.clone());
-                        tracing::debug!(
-                            new_version = %new_version_id,
-                            transform_id = %transform_id,
-                            "Linked pending migration to new schema version"
-                        );
-                    }
-                }
-            }
-        }
-
-        // Go compatibility: respect explicit IsActive=false in the patch, otherwise default to true.
-        // When IsActive was explicitly set to false in the patch, preserve it.
-        // When the new version is inactive, keep the old version active.
-        if !is_active_explicitly_set {
-            new_schema.is_active = true;
-        }
-
-        // Create old schema copy for storage. If new schema is active, mark old as inactive.
-        // If new schema is inactive (explicit IsActive=false), old version stays active.
-        let mut old_schema_inactive = old_schema.clone();
-        if new_schema.is_active {
-            old_schema_inactive.is_active = false;
-        }
-
-        tracing::info!(
-            collection = %collection_name,
-            old_version = %old_version_id,
-            new_version = %new_version_id,
-            field_count = new_schema.fields.len(),
-            "Creating new schema version"
-        );
-
-        // Begin transaction to store all version data
-        let txn = self.new_txn(false).await?;
-
-        // Prepare serialized data before getting systemstore reference
-        let old_version_key = CollectionKey::new(&old_version_id);
-        let old_version_data = serde_json::to_vec(&old_schema_inactive).map_err(|e| {
-            Error::Serialization(format!(
-                "failed to serialize old schema version '{}': {}",
-                old_version_id, e
-            ))
-        })?;
-        let new_version_key = CollectionKey::new(&new_version_id);
-        let new_version_data = serde_json::to_vec(&new_schema).map_err(|e| {
-            Error::Serialization(format!(
-                "failed to serialize new schema version '{}': {}",
-                new_version_id, e
-            ))
-        })?;
-        let name_key = CollectionNameKey::new(&actual_name);
-        let version_index_key = CollectionVersionKey::new(&collection_id, &new_version_id);
-        let old_version_index_key = CollectionVersionKey::new(&collection_id, &old_version_id);
-
-        // Perform all writes in a scoped block so systemstore reference is dropped
-        {
-            let systemstore = txn.systemstore()?;
-
-            // 1. Store old version at /collection/id/{old_version_id} with is_active = false
-            systemstore
-                .set(&old_version_key.bytes(), &old_version_data)
-                .await
-                .map_err(Error::Storage)?;
-
-            // 2. Store new version at /collection/id/{new_version_id}
-            systemstore
-                .set(&new_version_key.bytes(), &new_version_data)
-                .await
-                .map_err(Error::Storage)?;
-
-            // 3. Update /collection/name/{name} - only point to new version if it's active.
-            // If new version is inactive, keep name pointing to old version (which stays active).
-            if new_schema.is_active {
-                systemstore
-                    .set(&name_key.bytes(), new_version_id.as_bytes())
-                    .await
-                    .map_err(Error::Storage)?;
-            }
-
-            // 4. Add version index at /collection/version/{collection_id}/{new_version_id}
-            systemstore
-                .set(&version_index_key.bytes(), b"1")
-                .await
-                .map_err(Error::Storage)?;
-
-            // 5. Also ensure old version is in the version index (may already exist)
-            systemstore
-                .set(&old_version_index_key.bytes(), b"1")
-                .await
-                .map_err(Error::Storage)?;
-        } // systemstore reference dropped here
-
-        txn.commit().await?;
-
-        // Clean up any pending migration that was linked to this version
-        {
-            let mut pending = self.pending_migrations.write().map_err(|e| {
-                tracing::error!(error = ?e, "Pending migrations lock poisoned during cleanup");
-                Error::CacheUpdateFailedAfterCommit(collection_name.to_string())
-            })?;
-            pending.remove(&new_version_id);
-        }
-
-        // Update cache based on which version is active
-        let mut cache = self.collections.write().map_err(|e| {
-            tracing::error!(
-                error = ?e,
-                collection_name = %collection_name,
-                "Collection cache lock poisoned during patch_collection update"
-            );
-            Error::CacheUpdateFailedAfterCommit(collection_name.to_string())
-        })?;
-        if new_schema.is_active {
-            // New version is active - cache it under the actual collection name
-            // (not collection_name, which might be a version_id for branching patches)
-            cache.insert(actual_name.clone(), Collection::new(new_schema.clone()));
-        }
-        // If new version is inactive, old version stays in cache (already there)
-
-        Ok(new_schema)
-    }
-
-    /// Generate a version ID (CID) from schema content during patching.
-    ///
-    /// Matches Go DefraDB's saveBlocks() behavior:
-    /// - Existing fields (present in old_schema) are SKIPPED entirely
-    /// - Only NEW fields get CIDs generated with priority=1 (empty headstore)
-    /// - The collection block gets priority=version_depth+1, heads=[old_version_CID],
-    ///   and links containing only new field CIDs
-    fn generate_patch_version_id(
-        schema: &mut CollectionVersion,
-        old_schema: &CollectionVersion,
-        version_depth: u64,
-    ) -> String {
-        use cid::Cid;
-        use sha2::{Digest, Sha256};
-        use std::str::FromStr;
-
-        let collection_priority = version_depth + 1;
-
-        // Build set of old field names for detecting which fields are new
-        let old_field_names: std::collections::HashSet<&str> = old_schema
-            .fields
-            .iter()
-            .filter(|f| !f.id.is_empty())
-            .map(|f| f.name.as_str())
-            .collect();
-
-        // Go's saveBlocks skips fields that already have a FieldID (existing fields).
-        // Only NEW fields (not in old_schema) get CID generation and DAGLink inclusion.
-        // Go also skips secondary relation fields (those have empty FieldID in old schema too,
-        // but Delta returns hasFieldChanged=false for them).
-        let new_field_indices: Vec<usize> = {
-            let mut indices: Vec<usize> = schema
-                .fields
-                .iter()
-                .enumerate()
-                .filter(|(_, f)| {
-                    // New field: not in old schema and not a secondary relation
-                    // (secondary relations have relation_name set and is_primary=false)
-                    let is_new = !old_field_names.contains(f.name.as_str());
-                    let is_secondary_relation = f.relation_name.is_some() && !f.is_primary;
-                    is_new && !is_secondary_relation
-                })
-                .map(|(i, _)| i)
-                .collect();
-            // Sort: _docID first, then alphabetically
-            indices.sort_by(|&a, &b| {
-                let fa = &schema.fields[a];
-                let fb = &schema.fields[b];
-                if fa.name == "_docID" {
-                    std::cmp::Ordering::Less
-                } else if fb.name == "_docID" {
-                    std::cmp::Ordering::Greater
-                } else {
-                    fa.name.cmp(&fb.name)
-                }
-            });
-            indices
-        };
-
-        // Generate CIDs only for NEW fields with priority=1 (matching Go's empty headstore)
-        let mut field_cids: Vec<Cid> = Vec::new();
-        for &idx in &new_field_indices {
-            let field = &schema.fields[idx];
-            match schema::generate_field_cid_with_priority(field, 1) {
-                Ok(cid) => {
-                    schema.fields[idx].id = cid.to_string();
-                    field_cids.push(cid);
-                }
-                Err(_e) => {}
-            }
-        }
-
-        // Generate collection CID with old version as head.
-        // Go's Delta only includes name when it changed. For field-only patches, name=None.
-        let name_changed = schema.name != old_schema.name;
-        let collection_name = if name_changed {
-            Some(schema.name.as_str())
-        } else {
-            None
-        };
-        let old_version_cid = Cid::from_str(&old_schema.version_id).ok();
-        let collection_heads: Vec<Cid> = old_version_cid.into_iter().collect();
-        match schema::generate_collection_cid_full(
-            collection_name,
-            &field_cids,
-            collection_priority,
-            &collection_heads,
-        ) {
-            Ok(cid) => cid.to_string(),
-            Err(_) => {
-                // Fallback to simple hash if CID generation fails
-                let mut hasher = Sha256::new();
-                hasher.update(b"version:");
-                hasher.update(schema.name.as_bytes());
-                for field in &schema.fields {
-                    hasher.update(field.name.as_bytes());
-                    hasher.update(field.id.as_bytes());
-                }
-                let hash = hasher.finalize();
-                format!(
-                    "v{:x}",
-                    &hash[..8].iter().fold(0u64, |acc, &b| (acc << 8) | b as u64)
-                )
-            }
-        }
-    }
-
-    /// Strip collection name or version ID prefix from a path.
-    ///
-    /// Handles both the collection_name prefix (e.g., "/Users/") and the
-    /// actual_name prefix (when looked up by version ID, the passed-in name
-    /// differs from the real collection name).
-    /// Validate a Kind value in a patch field addition.
-    /// Returns error if the Kind is an unsupported numeric value or unknown string.
-    fn validate_patch_field_kind(
-        kind_val: &serde_json::Value,
-        field_name: &str,
-        known_collections: &[String],
-    ) -> Result<()> {
-        match kind_val {
-            serde_json::Value::Number(n) => {
-                let kind_num = n.as_u64().unwrap_or(0) as u8;
-                // Valid numeric kinds: 1-14, 18-22 (0 is None, only for internal _docID)
-                let valid = matches!(
-                    kind_num,
-                    1..=14 | 18..=22
-                );
-                if !valid {
-                    return Err(Error::InvalidPatch(format!(
-                        "no type found for given name. Type: {}",
-                        kind_num
-                    )));
-                }
-                Ok(())
-            }
-            serde_json::Value::String(s) => {
-                // Known string kinds
-                let known = matches!(
-                    s.as_str(),
-                    "ID" | "Boolean"
-                        | "Int"
-                        | "DateTime"
-                        | "Float"
-                        | "Float64"
-                        | "Float32"
-                        | "String"
-                        | "Blob"
-                        | "JSON"
-                        | "[Boolean]"
-                        | "[Boolean!]"
-                        | "[Int]"
-                        | "[Int!]"
-                        | "[Float]"
-                        | "[Float64]"
-                        | "[Float!]"
-                        | "[Float64!]"
-                        | "[Float32]"
-                        | "[Float32!]"
-                        | "[String]"
-                        | "[String!]"
-                        | "Self"
-                        | "[Self]"
-                );
-                if !known {
-                    // Could be a collection name reference (e.g., "Users", "[Users]").
-                    let ref_name = s
-                        .strip_prefix('[')
-                        .and_then(|s| s.strip_suffix(']'))
-                        .unwrap_or(s.as_str());
-                    if !known_collections.iter().any(|c| c == ref_name) {
-                        return Err(Error::InvalidPatch(format!(
-                            "no type found for given name. Field: {}, Kind: {}",
-                            field_name, s
-                        )));
-                    }
-                }
-                Ok(())
-            }
-            _ => Ok(()),
-        }
-    }
-
-    fn strip_collection_prefix(path: &str, prefixes: &[String]) -> String {
-        // Go DefraDB accepts paths with or without leading '/'.
-        // Try each prefix (collection name, actual name, version ID) in order.
-        for prefix in prefixes {
-            let no_slash_prefix = prefix.trim_start_matches('/');
-
-            if path.starts_with(prefix.as_str()) {
-                return format!("/{}", &path[prefix.len()..]);
-            }
-            if path.starts_with(no_slash_prefix) {
-                return format!("/{}", &path[no_slash_prefix.len()..]);
-            }
-            // Exact match without trailing slash (collection-level operations).
-            // E.g., path="/Users" with prefix="/Users/" → "/"
-            let exact = prefix.trim_end_matches('/');
-            let exact_no_slash = exact.trim_start_matches('/');
-            if path == exact || path == exact_no_slash {
-                return "/".to_string();
-            }
-        }
-        path.to_string()
-    }
-
-    /// Handle patches targeting a collection that doesn't exist by name or version ID.
-    ///
-    /// This handles several cases:
-    /// 1. Schema field names (EncryptedIndexes, Indexes, etc.) - produce JSON patch errors
-    ///    because these fields don't exist in Go's JSON representation
-    /// 2. Collection-level copy where the "path" targets a new collection name
-    ///    (e.g., copy from /Users to /Book) → returns "adding collections not supported"
-    /// 3. Collection-level move to a new name (no-op in Go) → finds source via "from"
-    ///    and returns the original schema unchanged
-    async fn handle_unknown_collection_patch(
-        &self,
-        collection_name: &str,
-        patch_ops: &serde_json::Value,
-    ) -> Result<CollectionVersion> {
-        // Schema field names that don't exist in Go's JSON representation
-        // When the "collection name" is actually one of these, produce Go-compatible
-        // JSON patch errors instead of "adding collections" errors.
-        const SCHEMA_FIELDS: &[&str] = &[
-            "EncryptedIndexes",
-            "VectorEmbeddings",
-            "Indexes",
-            "Fields",
-            "Policy",
-        ];
-
-        // Check if the "collection name" is actually a schema field
-        if SCHEMA_FIELDS.contains(&collection_name) {
-            if let serde_json::Value::Array(ops) = patch_ops {
-                for op in ops {
-                    let operation = op.get("op").and_then(|v| v.as_str());
-                    let value = op.get("value");
-
-                    match operation {
-                        Some("add") => {
-                            // For add with array value, Go produces unmarshal error
-                            if value.map(|v| v.is_array()).unwrap_or(false) {
-                                return Err(Error::InvalidPatch(
-                                    "cannot unmarshal array into Go value".to_string(),
-                                ));
-                            }
-                            return Err(Error::InvalidPatch(
-                                "cannot unmarshal array into Go value".to_string(),
-                            ));
-                        }
-                        Some("remove") => {
-                            return Err(Error::InvalidPatch(
-                                "unable to remove nonexistent key".to_string(),
-                            ));
-                        }
-                        Some("replace") => {
-                            return Err(Error::InvalidPatch("doc is missing key".to_string()));
-                        }
-                        _ => {}
-                    }
-                }
-            }
-        }
-
-        // Try to extract the actual collection name from the patch value's Name field.
-        // This handles cases like path "/-" where the collection name in the path is "-"
-        // but the actual name is in the value object.
-        let effective_name = if let serde_json::Value::Array(ops) = patch_ops {
-            ops.iter()
-                .find_map(|op| {
-                    op.get("value")
-                        .and_then(|v| v.get("Name"))
-                        .and_then(|n| n.as_str())
-                        .map(String::from)
-                })
-                .unwrap_or_else(|| collection_name.to_string())
-        } else {
-            collection_name.to_string()
-        };
-
-        if let serde_json::Value::Array(ops) = patch_ops {
-            // Look for move/copy operations to determine if this is a routing issue
-            for op in ops {
-                let operation = op.get("op").and_then(|v| v.as_str());
-                let from_raw = op.get("from").and_then(|v| v.as_str());
-
-                match operation {
-                    Some("add") => {
-                        // Check if the patch targets a sub-path within the unknown collection.
-                        // If so, it's "doc is missing path" (the collection doesn't exist).
-                        // If it targets the root (no sub-path), it's "adding collections not supported".
-                        let raw_path = op.get("path").and_then(|v| v.as_str()).unwrap_or("");
-                        let trimmed = raw_path.trim_start_matches('/');
-                        let has_subpath = trimmed.contains('/');
-                        if has_subpath {
-                            return Err(Error::InvalidPatch(
-                                "add operation does not apply: doc is missing path".to_string(),
-                            ));
-                        }
-                        return Err(Error::InvalidPatch(format!(
-                            "adding collections via patch is not supported. Name: {}",
-                            effective_name,
-                        )));
-                    }
-                    Some("copy") | Some("replace") => {
-                        return Err(Error::InvalidPatch(format!(
-                            "adding collections via patch is not supported. Name: {}",
-                            effective_name,
-                        )));
-                    }
-                    Some("move") => {
-                        // Collection-level move is a no-op - find source and return unchanged
-                        if let Some(from) = from_raw {
-                            let source_name =
-                                from.trim_start_matches('/').split('/').next().unwrap_or("");
-                            if let Some(source_col) = self.get_collection(source_name)? {
-                                return Ok(source_col.schema().clone());
-                            }
-                        }
-                    }
-                    _ => {}
-                }
-            }
-        }
-
-        // No recognized operation - adding collections via patch is not supported
-        Err(Error::InvalidPatch(format!(
-            "adding collections via patch is not supported. Name: {}",
-            effective_name,
-        )))
-    }
-
-    /// Helper: Set a value at a JSON pointer path.
-    fn json_pointer_set(
-        json: &mut serde_json::Value,
-        path: &str,
-        value: serde_json::Value,
-    ) -> Result<()> {
-        // Convert JSON pointer to path segments
-        let segments: Vec<&str> = path.split('/').filter(|s| !s.is_empty()).collect();
-        if segments.is_empty() {
-            return Err(Error::InvalidPatch("empty path".to_string()));
-        }
-
-        let mut current = json;
-        for (i, segment) in segments.iter().enumerate() {
-            if i == segments.len() - 1 {
-                // Last segment - set the value
-                match current {
-                    serde_json::Value::Object(map) => {
-                        map.insert(segment.to_string(), value);
-                        return Ok(());
-                    }
-                    serde_json::Value::Array(arr) => {
-                        // JSON Pointer uses "-" to mean "append to end of array"
-                        if *segment == "-" {
-                            arr.push(value);
-                        } else {
-                            let idx: usize = segment.parse().map_err(|_| {
-                                Error::InvalidPatch(format!("invalid array index: {}", segment))
-                            })?;
-                            if idx >= arr.len() {
-                                arr.push(value);
-                            } else {
-                                arr[idx] = value;
-                            }
-                        }
-                        return Ok(());
-                    }
-                    _ => {
-                        return Err(Error::InvalidPatch(format!(
-                            "cannot set value at path {}",
-                            path
-                        )));
-                    }
-                }
-            } else {
-                // Navigate to the next level
-                match current {
-                    serde_json::Value::Object(map) => {
-                        current = map.get_mut(*segment).ok_or_else(|| {
-                            Error::InvalidPatch(format!("path not found: {}", path))
-                        })?;
-                    }
-                    serde_json::Value::Array(arr) => {
-                        let idx: usize = segment.parse().map_err(|_| {
-                            Error::InvalidPatch(format!("invalid array index: {}", segment))
-                        })?;
-                        current = arr.get_mut(idx).ok_or_else(|| {
-                            Error::InvalidPatch(format!("path not found: {}", path))
-                        })?;
-                    }
-                    _ => {
-                        return Err(Error::InvalidPatch(format!(
-                            "cannot navigate path: {}",
-                            path
-                        )));
-                    }
-                }
-            }
-        }
-
-        Err(Error::InvalidPatch("failed to set value".to_string()))
-    }
-
-    /// Helper: Remove a value at a JSON pointer path.
-    fn json_pointer_remove(json: &mut serde_json::Value, path: &str) -> Result<()> {
-        let segments: Vec<&str> = path.split('/').filter(|s| !s.is_empty()).collect();
-        if segments.is_empty() {
-            return Err(Error::InvalidPatch("empty path".to_string()));
-        }
-
-        let mut current = json;
-        for (i, segment) in segments.iter().enumerate() {
-            if i == segments.len() - 1 {
-                // Last segment - remove the value
-                match current {
-                    serde_json::Value::Object(map) => {
-                        map.remove(*segment);
-                        return Ok(());
-                    }
-                    serde_json::Value::Array(arr) => {
-                        let idx: usize = segment.parse().map_err(|_| {
-                            Error::InvalidPatch(format!("invalid array index: {}", segment))
-                        })?;
-                        if idx < arr.len() {
-                            arr.remove(idx);
-                        }
-                        return Ok(());
-                    }
-                    _ => {
-                        return Err(Error::InvalidPatch(format!(
-                            "cannot remove value at path {}",
-                            path
-                        )));
-                    }
-                }
-            } else {
-                // Navigate to the next level
-                match current {
-                    serde_json::Value::Object(map) => {
-                        current = map.get_mut(*segment).ok_or_else(|| {
-                            Error::InvalidPatch(format!("path not found: {}", path))
-                        })?;
-                    }
-                    serde_json::Value::Array(arr) => {
-                        let idx: usize = segment.parse().map_err(|_| {
-                            Error::InvalidPatch(format!("invalid array index: {}", segment))
-                        })?;
-                        current = arr.get_mut(idx).ok_or_else(|| {
-                            Error::InvalidPatch(format!("path not found: {}", path))
-                        })?;
-                    }
-                    _ => {
-                        return Err(Error::InvalidPatch(format!(
-                            "cannot navigate path: {}",
-                            path
-                        )));
-                    }
-                }
-            }
-        }
-
-        Err(Error::InvalidPatch("failed to remove value".to_string()))
-    }
-
-    /// Extract a field name from a path like `/Fields/email` or `/Fields/email/Name`.
-    /// Returns None if the segment after /Fields/ is numeric, "-", or /Fields/ isn't present.
-    fn extract_field_name_from_path(path: &str) -> Option<String> {
-        let segments: Vec<&str> = path.split('/').collect();
-        for (i, seg) in segments.iter().enumerate() {
-            if *seg == "Fields" && i + 1 < segments.len() {
-                let next = segments[i + 1];
-                if next.parse::<usize>().is_ok() || next == "-" {
-                    return None;
-                }
-                return Some(next.to_string());
-            }
-        }
-        None
-    }
-
-    /// Helper: Get a value at a JSON pointer path (for test operation).
-    fn json_pointer_get(json: &serde_json::Value, path: &str) -> Option<serde_json::Value> {
-        let segments: Vec<&str> = path.split('/').filter(|s| !s.is_empty()).collect();
-        if segments.is_empty() {
-            return None;
-        }
-
-        let mut current = json;
-        for segment in segments.iter() {
-            match current {
-                serde_json::Value::Object(map) => {
-                    current = map.get(*segment)?;
-                }
-                serde_json::Value::Array(arr) => {
-                    let idx: usize = segment.parse().ok()?;
-                    current = arr.get(idx)?;
-                }
-                _ => return None,
-            }
-        }
-        Some(current.clone())
-    }
-
-    /// Helper: Substitute field names for indices in paths like /Fields/<name>
-    /// Go DefraDB allows using field names as array indices in patches.
-    fn substitute_field_name_in_path(path: &str, schema_json: &serde_json::Value) -> String {
-        // Check if path contains /Fields/ followed by a non-numeric segment
-        if !path.contains("/Fields/") {
-            return path.to_string();
-        }
-
-        let segments: Vec<&str> = path.split('/').collect();
-        let mut result_segments: Vec<String> = Vec::new();
-
-        let mut i = 0;
-        while i < segments.len() {
-            let segment = segments[i];
-
-            if segment == "Fields" && i + 1 < segments.len() {
-                result_segments.push("Fields".to_string());
-                i += 1;
-
-                let next_segment = segments[i];
-                // Check if next segment is a number (already an index)
-                if next_segment.parse::<usize>().is_ok() || next_segment == "-" {
-                    result_segments.push(next_segment.to_string());
-                } else {
-                    // It's a field name - look up the index in the existing Fields array
-                    if let Some(fields) = schema_json.get("Fields").and_then(|f| f.as_array()) {
-                        let mut found = false;
-                        for (idx, field) in fields.iter().enumerate() {
-                            if let Some(name) = field.get("Name").and_then(|n| n.as_str()) {
-                                if name == next_segment {
-                                    result_segments.push(idx.to_string());
-                                    found = true;
-                                    break;
-                                }
-                            }
-                        }
-                        if !found {
-                            // Field name not found in existing fields - treat as append
-                            // (Go interprets unknown field names as new field additions)
-                            result_segments.push("-".to_string());
-                        }
-                    } else {
-                        result_segments.push("-".to_string());
-                    }
-                }
-            } else {
-                result_segments.push(segment.to_string());
-            }
-            i += 1;
-        }
-
-        result_segments.join("/")
     }
 
     /// Extract a Merkle proof from the blockstore.
@@ -3843,76 +1534,4 @@ fn extract_last_path_segment_str(key: &[u8]) -> Option<String> {
         }
     }
     None
-}
-
-/// Convert a JSON value to a native NormalValue based on the field's schema type.
-///
-/// When documents are migrated through lens transforms, they come back as JSON values.
-/// This function converts them to the appropriate native type (Int, Float, String, etc.)
-/// based on the field's declared type in the schema.
-fn json_to_native_value(
-    value: &serde_json::Value,
-    field_name: &str,
-    schema: &schema::CollectionVersion,
-) -> document::NormalValue {
-    // Handle null values
-    if value.is_null() {
-        return document::NormalValue::Null;
-    }
-
-    // Find the field definition in the schema
-    let field_kind = schema
-        .fields
-        .iter()
-        .find(|f| f.name == field_name)
-        .map(|f| &f.kind);
-
-    match field_kind {
-        Some(FieldKind::Scalar(scalar)) => match scalar {
-            ScalarKind::Int => {
-                if let Some(n) = value.as_i64() {
-                    return document::NormalValue::Int(n);
-                }
-            }
-            ScalarKind::Float64 => {
-                if let Some(n) = value.as_f64() {
-                    return document::NormalValue::Float64(n);
-                }
-            }
-            ScalarKind::Float32 => {
-                if let Some(n) = value.as_f64() {
-                    return document::NormalValue::Float32(n as f32);
-                }
-            }
-            ScalarKind::Bool => {
-                if let Some(b) = value.as_bool() {
-                    return document::NormalValue::Bool(b);
-                }
-            }
-            ScalarKind::String | ScalarKind::DocID => {
-                if let Some(s) = value.as_str() {
-                    return document::NormalValue::String(s.to_string());
-                }
-            }
-            ScalarKind::Blob => {
-                // Blobs may be base64 encoded strings in JSON
-                if let Some(s) = value.as_str() {
-                    return document::NormalValue::Bytes(s.as_bytes().to_vec());
-                }
-            }
-            ScalarKind::DateTime => {
-                // DateTime as string - keep as string for now, the document layer handles parsing
-                if let Some(s) = value.as_str() {
-                    return document::NormalValue::String(s.to_string());
-                }
-            }
-            ScalarKind::Json | ScalarKind::None => {
-                // Keep as JSON
-            }
-        },
-        _ => {}
-    }
-
-    // Fallback: keep as JSON (this preserves the original behavior for unknown types)
-    document::NormalValue::Json(value.clone())
 }
